Deploy/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
phonenumbers==8.13.37
redis==5.0.4
PyJWT==2.8.0
python-multipart==0.0.9
